risk-engine/pyproject.toml
[project]
name = "kinetix-risk"
version = "0.1.0"
description = "Python risk engine for Kinetix — VaR, Monte Carlo, Greeks, ML models"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "grpcio>=1.60",
    "protobuf>=4.25",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "grpcio-tools>=1.60",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kinetix_risk"]
